Reorganise test files and add one specific to GC
