+++ syft/frameworks/torch/tensors/__init__.py
@@ -1,4 +1,4 @@
-from .log import LoggingTensor  # noqa: F401
+from .logging import LoggingTensor  # noqa: F401
 from .pointer import PointerTensor  # noqa: F401
 from .abstract import AbstractTensor  # noqa: F401
 from .native import TorchTensor  # noqa: F401
